requirements.txt
qrcode==5.3
setuptools==17.0
Pillow==4.3.0
